projects/services/ingestion/web-crawl/examples/requirements.txt
# Requirements for running examples locally
# Install with: pip install -r requirements.txt

# Core dependencies
crawl4ai>=0.7.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Database persistence
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
